fix 0.1

+++ marishin.leonid/P3/inputstring.cpp
@@ -32,7 +32,7 @@ char* marishin::inputString(std::istream& input)
         delete[] buffer;
         throw ba;
       }
-      for (int i = 0; i < currentIndex; ++i)
+      for (size_t i = 0; i < currentIndex; ++i)
       {
         newBuffer[i] = buffer[i];
       }
